Merge request from develop into release

See merge request s08-final/S08P31C101!692

+++ backend/user-service/src/main/java/com/example/userservice/dto/request/user/LocationRequestDto.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class LocationRequestDto {
-    private Float lat;
-    private Float lng;
+    private Double lat;
+    private Double lng;
     private Boolean moving;
 }
